\name{crossval-internal}

\alias{group.samples}
\alias{get.fold}

\alias{crossval-internal}


\title{Internal crossval Functions}
\description{
  Internal crossval functions.
}
\note{
  These are not to be called by the user (or in some cases are just
  waiting for proper documentation to be written).
}
\keyword{internal}
