\encoding{latin1}
\name{crossval-package}
\alias{crossval-package}
\docType{package}
\title{The crossval Package}

\description{
The "crossval" package implements generic functions for performing 
  cross validation and for computing diagnostic errors.
}

\author{Korbinian Strimmer (\url{https://strimmerlab.github.io/})}
\references{
  
  Website: \url{https://cran.r-project.org/package=crossval}
}
\keyword{multivariate}
\seealso{
\code{\link{crossval}}, \code{\link{confusionMatrix}}, \code{\link{diagnosticErrors}}.
}
