\encoding{latin1}
\name{crossval-package}
\alias{crossval-package}
\docType{package}
\title{The crossval Package}

\description{
The "crossval" package implements generic functions for performing 
  cross validation and for computing diagnostic errors.
}

\author{Korbinian Strimmer (\url{http://strimmerlab.org/})}
\references{
  
  Website: \url{http://cran.r-project.org/web/packages/crossval/}
}
\keyword{multivariate}
\seealso{
\code{\link{crossval}}, \code{\link{confusionMatrix}}, \code{\link{diagnosticErrors}}.
}
